src/main/webapp/app/entities/book-status/delete/book-status-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { BookStatusService } from '../service/book-status.service';

import { BookStatusDeleteDialogComponent } from './book-status-delete-dialog.component';

describe('BookStatus Management Delete Component', () => {
  let comp: BookStatusDeleteDialogComponent;
  let fixture: ComponentFixture<BookStatusDeleteDialogComponent>;
  let service: BookStatusService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, BookStatusDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(BookStatusDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(BookStatusDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(BookStatusService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
